+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Add action for cloning a unit with subordinates
 - Simplified standard identity selection.
 - Add zooming and panning to ORBAT charts.
 - Add vitepress-powered documentation https://docs.orbat-mapper.app (work in progress).